creando dtos y enviado datos de prueba a la bd

+++ src/products/entities/product.entity.ts
@@ -11,7 +11,7 @@ export class Product {
     })
     title: string
 
-    @Column('numeric', { default: 0 })
+    @Column('float', { default: 0 })
     price: number
 
     @Column({ type: 'varchar', nullable: true })
